use appropriate index types in transit shortest paths containers

+++ src/ShortestPath/Transit/TransitShortestPathContainer.cpp
@@ -6,18 +6,18 @@
 #include "TransitShortestPathContainer.h"
 #include "TransitStateContainer.h"
 
-void TransitShortestPathContainer::addShortestPathCollection(int startNodeIndex,
-                                                             const std::pair<int, std::vector<TransitShortestPath>>& shortestPathList) {
+void TransitShortestPathContainer::addShortestPathCollection(size_t startNodeIndex,
+                                                             const std::pair<size_t , std::vector<TransitShortestPath>>& shortestPathList) {
     container[startNodeIndex].emplace_back(shortestPathList);
 }
 
-void TransitShortestPathContainer::addShortestPathCollection(int startNodeIndex, int startingInstant, int graphSize,
+void TransitShortestPathContainer::addShortestPathCollection(size_t startNodeIndex, uint startingInstant, size_t graphSize,
                                                              const TransitStateContainer& algorithmResultStates) {
     std::vector<TransitShortestPath> shortestPathList;
     shortestPathList.reserve(graphSize);
 
     //Convert states to shortest paths and add to collection
-    for(int i = 0; i < graphSize; ++i) {
+    for(size_t i = 0; i < graphSize; ++i) {
         shortestPathList.emplace_back(algorithmResultStates.getBestSolution(i));
     }
 
@@ -25,16 +25,16 @@ void TransitShortestPathContainer::addShortestPathCollection(int startNodeIndex,
     container[startNodeIndex].emplace_back(startingInstant, shortestPathList);
 }
 
-std::pair<int, std::vector<TransitShortestPath>> *
-TransitShortestPathContainer::getShortestPathsFrom(int startNodeIndex, int earliestStartInstant) {
+std::pair<size_t, std::vector<TransitShortestPath>> *
+TransitShortestPathContainer::getShortestPathsFrom(size_t startNodeIndex, uint earliestStartInstant) {
     const auto& iterator = std::ranges::lower_bound(container[startNodeIndex],
-                                            std::pair<int, std::vector<TransitShortestPath>>(earliestStartInstant, {}));
+                                            std::pair<size_t , std::vector<TransitShortestPath>>(earliestStartInstant, {}));
 
     return iterator.base();
 }
 
 TransitShortestPath
-TransitShortestPathContainer::getShortestPathToYFromTime(int startNodeIndex, int earliestStartInstant, int goalNode) {
+TransitShortestPathContainer::getShortestPathToYFromTime(size_t startNodeIndex, uint earliestStartInstant, size_t goalNode) {
     auto [startNode, shortestPathsVector] = *getShortestPathsFrom(startNodeIndex, earliestStartInstant);
     for(const auto& shortestPath : shortestPathsVector) {
         if(shortestPath.getArrivalNode() == goalNode) {
